Move configs to separate files
