+++ internal/db/provider.go
@@ -8,13 +8,6 @@ import (
 	"github.com/jmoiron/sqlx"
 )
 
-type DbConfig struct {
-	Host     string `yaml:"host" validate:"required"`
-	Name     string `yaml:"name" validate:"required"`
-	User     string `yaml:"user" validate:"required"`
-	Password string `yaml:"password" validate:"required"`
-}
-
 type Provider struct {
 	*sqlx.DB
 }
