+++ internal/server/server.go
@@ -2,18 +2,9 @@ package server
 
 import (
 	"net/http"
-	"time"
 
 	"github.com/gorilla/mux"
 )
-
-type ServerConfig struct {
-	Host             *string        `yaml:"host"`
-	ResourceTimeout  *time.Duration `yaml:"resource-timeout"`
-	AnalyticsTimeout *time.Duration `yaml:"analytics-timeout"`
-	ReadTimeout      *time.Duration `yaml:"read-timeout"`
-	WriteTimeout     *time.Duration `yaml:"write-timeout"`
-}
 
 type Server struct {
 	server *http.Server
